pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "email-validator>=2.2.0",
    "etl>=0.0.1",
    "flask>=3.1.0",
    "flask-sqlalchemy>=3.1.1",
    "geopandas>=1.0.1",
    "gunicorn>=23.0.0",
    "pandas>=2.2.3",
    "psutil>=7.0.0",
    "psycopg2-binary>=2.9.10",
    "pyodbc>=5.2.0",
    "python-dotenv>=1.1.0",
    "shapely>=2.1.0",
]
